Fix code

+++ ports/cudnn/master/linux.sh
@@ -3,11 +3,13 @@
 #   cp -f -r /usr/lib/libcudnn* $INSTALL_PREFIX/cudnn/lib
 # else
 if ! command -v pacman; then
-  cd $SRC_DIR
-  cd cudnn*
-  rm -rf $INSTALL_PREFIX/cudnn
-  mkdir -p $INSTALL_PREFIX/cudnn
-  cp -f -r include $INSTALL_PREFIX/cudnn
-  cp -f -r lib $INSTALL_PREFIX/cudnn
-  rm -rf *
+  if ! command -v apt; then
+    cd $SRC_DIR
+    cd cudnn*
+    rm -rf $INSTALL_PREFIX/cudnn
+    mkdir -p $INSTALL_PREFIX/cudnn
+    cp -f -r include $INSTALL_PREFIX/cudnn
+    cp -f -r lib $INSTALL_PREFIX/cudnn
+    rm -rf *
+  fi
 fi
